example-agent/pyproject.toml
[project]
name = "example-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "parlant>=3.0.2",
    "python-dotenv>=1.1.1",
]
